Fixed removal of default phpldapadmin apache2 config file
In earlier Debian/Ubuntu versions the default apache2 config for phpldapadmin
was stored in /etc/apache2/conf.d. This directory no longer exists on Ubuntu
14.04, and the phpldapadmin apache2 config is supposed to be in
/etc/apache2/conf-enabled, and should be removed from there instead. That said,
due to a bug in the Ubuntu package the config file is still being copied to
conf.d (which does not exist):

<https://bugs.launchpad.net/ubuntu/+source/phpldapadmin/+bug/1297225>

Note that this commit depends on a new variable, $::apache2::params::conf_d_dir,
so you will need to update the apache2 puppet module too.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>