Worked around critical problems in the Ubuntu 14.04 package
The phpldapadmin sources in Ubuntu 14.04 (1.2.2-5ubuntu1) are broken in many
ways:

<https://bugs.launchpad.net/ubuntu/+source/phpldapadmin/+bug/1289237>
<https://bugs.launchpad.net/ubuntu/+source/phpldapadmin/+bug/1319144>

Added a patch to workaround these issues on Ubuntu 14.04:

<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=710854>

Tests are in place to ensure the patch will not be applied if the sources have
been fixed already. This means future phpldapadmin package upgrades will not
break this module.

Also appended .conf to the phpldapadmin VirtualHost filename, so that the
default apache2.conf on Ubuntu 14.04 loads it properly. This change is not
isolated to Ubuntu 14.04 but should be perfectly safe. Old VirtualHost files
installed by this module are removed by ldapadmin::absent.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>